requirements.txt
langchain-openai~=0.3.3
langchain~=0.3.17
langchain-community~=0.3.16
langchain-core~=0.3.33
wikipedia
youtube-search
python-dotenv~=1.0.1